Extract Rc Drop implementation to a cycle module
